test: keep Pi RPC lifecycle coverage in unit CI
